Rename changeset to a more generic name
